Pull up revision 1.3 (requested by tron in ticket #864):   Fix build problem on LP64 platforms.

+++ xfree/xc/extras/Xpm/lib/Attrib.c
@@ -52,7 +52,7 @@ CreateOldColorTable(ct, ncolors, oldct)
     XpmColor **colorTable, **color;
     int a;
 
-    if (ncolors >= SIZE_MAX / sizeof(XpmColor *)) 
+    if (ncolors >= UINT_MAX / sizeof(XpmColor *)) 
 	return XpmNoMemory;
 
     colorTable = (XpmColor **) XpmMalloc(ncolors * sizeof(XpmColor *));
